Set _dynamo.config.capture_func_transforms=False
Due to not all tests in the Dynamo shard actually running in CI, we've
started to bitrot on this implementation. Since our plan is to trace
into the functorch implementations instead of construct a HOP
(which is what capture_func_transforms=True does), let's turn off this
config by default.

Test Plan:
- Tests

ghstack-source-id: e05770ff42d8ff62e2037d2dde5a7a5434daf07b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115267