[export] optimize unflattener
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115364

Unflattening was slow on the APS FM model (which has thousands of nn.EmbeddingBag modules).

Quick glance at the profile shows 75% of time in unflattening was spent copying this node list, which is immutable and globally shared. So just passing it around as a tuple yields a 4x speedup lol.
ghstack-source-id: 209580532
@exported-using-ghexport

Differential Revision: [D51929775](https://our.internmc.facebook.com/intern/diff/D51929775/)