[Cutlass inductor backend] Cutlass GEMM size threshold
Cutlass backend GEMMs are comparatively expensive to compile. So they should only be applied to sufficiently large GEMMs. This small diff introduces a new torch._inductor.config option called "cuda.cutlass_backend_min_gemm_size" which introduces a threshold for the size of GEMM problems that the Cutlass backend will be considered for.

ghstack-source-id: 22f26e00679f3ecb2256b51fb74b8d7a278dad43
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113569